requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astropy
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.10.0
bokeh==2.4.3
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
cycler==0.11.0
Cython==0.29.28
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fonttools==4.29.1
gitdb==4.0.9
GitPython==3.1.27
gwdatafind==1.0.5
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.2
ipykernel==6.9.1
ipython
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jplephem==2.17
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lalsuite
ligo-segments
lscsoft-glue
Mako==1.1.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mpld3==0.5.7
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy
packaging==21.3
pandas
pandocfilters==1.5.0
parso==0.8.3
pegasus-wms.api==5.0.1
pegasus-wms.common==5.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
PyCBC
pycparser==2.21
pydeck==0.7.1
pyerfa==2.0.0.1
Pygments==2.11.2
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
pyRXP==3.0.1
python-dateutil==2.8.2
python-ligo-lw==1.8.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.27.1
scipy
semver==2.13.0
Send2Trash==1.8.0
sgp4==2.20
six==1.16.0
skyfield==1.42
smmap==5.0.0
soupsieve==2.3.1
stack-data==0.2.0
streamlit>=1.22.0
terminado==0.13.3
testpath==0.6.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.7.0
SoundFile==0.10.3.post1
wave
protobuf~=3.19.0

